env problème resolve

+++ BackEnd/server.js
@@ -37,29 +37,30 @@ app.get('/',(req,res)=>{
 
 app.get('/node/transfertAutorisation',(req,res)=>{
 
-    client.connect(err =>{
-        async function find(){
-            try {
+    let reponse;
 
-                const database = client.db("BigOne");
-                const enAttente = database.collection("enAttente");
-                const confirm = database.collection("confirm");
-
-                search = await enAttente.find({}).toArray();
-
-                transfert = await confirm.insertMany(search);
-
-                detruire = await enAttente.deleteMany({});
-
-            }
-            finally{
-                await client.close(); 
-            }}
-            find().catch();  
-    }); 
-
+    client.connect(async err => {
+        try {
+          const database = client.db("BigOne");
+          const enAttente = database.collection("enAttente");
+          const confirm = database.collection("confirm");
+      
+          const search = await enAttente.find({}).toArray();
+          if (search.length === 0) {
+            res.end("Il n'y a aucune demande");
+            return;
+          }
+      
+          const transfert = await confirm.insertMany(search);
+          const detruire = await enAttente.deleteMany({});
+          res.end("Autorisation accordée");
+        } catch (error) {
+          res.end("Une erreur s'est produite");
+        } finally {
+          await client.close();
+        }
+    });
     
-    res.send('<p id="autotrue">Autorisation accordée</p>');
 })
 
 
@@ -292,8 +293,6 @@ client.connect(err =>{
             const confirmCollection = database.collection("confirm");
             const tempsReelCollection = database.collection("tempsReel");
             const query = req.body;
-
-            console.log(query)
 
             const result = await confirmCollection.findOne(query);
